+++ src/main/java/it/gov/pagopa/rtd/ms/enrolledpaymentinstrument/infrastructure/persistence/repositories/EnrolledPaymentInstrumentDao.java
@@ -1,5 +1,6 @@
-package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.infrastructure.persistence.mongo.model;
+package it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.infrastructure.persistence.repositories;
 
+import it.gov.pagopa.rtd.ms.enrolledpaymentinstrument.infrastructure.persistence.mongo.model.EnrolledPaymentInstrumentEntity;
 import java.util.Optional;
 import org.springframework.data.mongodb.repository.MongoRepository;
